+++ src/main/java/com/commercehub/gradle/plugin/avro/DefaultAvroExtension.java
@@ -1,19 +1,28 @@
 package com.commercehub.gradle.plugin.avro;
 
+import org.apache.avro.compiler.specific.SpecificCompiler;
+import org.apache.avro.generic.GenericData;
+
+import java.nio.charset.Charset;
+
 public class DefaultAvroExtension implements AvroExtension {
-    private String encoding;
+    private String outputCharacterEncoding;
     private String stringType;
     private String fieldVisibility;
     private String templateDirectory;
     private boolean createSetters;
 
     @Override
-    public String getEncoding() {
-        return encoding;
+    public String getOutputCharacterEncoding() {
+        return outputCharacterEncoding;
     }
 
-    public void setEncoding(String encoding) {
-        this.encoding = encoding;
+    public void setOutputCharacterEncoding(String outputCharacterEncoding) {
+        this.outputCharacterEncoding = outputCharacterEncoding;
+    }
+
+    public void setOutputCharacterEncoding(Charset outputCharacterEncoding) {
+        setOutputCharacterEncoding(outputCharacterEncoding.name());
     }
 
     @Override
@@ -25,6 +34,10 @@ public class DefaultAvroExtension implements AvroExtension {
         this.stringType = stringType;
     }
 
+    public void setStringType(GenericData.StringType stringType) {
+        setStringType(stringType.name());
+    }
+
     @Override
     public String getFieldVisibility() {
         return fieldVisibility;
@@ -32,6 +45,10 @@ public class DefaultAvroExtension implements AvroExtension {
     
     public void setFieldVisibility(String fieldVisibility) {
         this.fieldVisibility = fieldVisibility;
+    }
+
+    public void setFieldVisibility(SpecificCompiler.FieldVisibility fieldVisibility) {
+        setFieldVisibility(fieldVisibility.name());
     }
 
     @Override
@@ -46,10 +63,6 @@ public class DefaultAvroExtension implements AvroExtension {
     @Override
     public boolean isCreateSetters() {
         return createSetters;
-    }
-
-    public void setCreateSetters(boolean createSetters) {
-        this.createSetters = createSetters;
     }
 
     public void setCreateSetters(String createSetters) {